@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 
-import { Observable } from 'rxjs/Observable';
-import { of } from 'rxjs/observable/of';
+import { Observable, of } from 'rxjs';
 import { map } from 'rxjs/operators';
+
 import { ICustomer } from '../../shared/interfaces';
 import { ClonerService } from './cloner.service';
 import { Subject, BehaviorSubject } from 'rxjs';
